1.18.1.1: Updated #1480

+++ source/res/scripts/client/gui/shared/close_confiramtor_helper.py
@@ -1,6 +1,8 @@
 # Python bytecode 2.7 (decompiled from Python 2.7)
 # Embedded file name: scripts/client/gui/shared/close_confiramtor_helper.py
 import adisp
+from collections import namedtuple
+from gui.game_control.links import URLMacros
 from wg_async import wg_async, wg_await
 from gui.Scaleform.daapi.settings.views import VIEW_ALIAS
 from gui.impl.gen import R
@@ -8,26 +10,33 @@ from gui.prb_control.entities.base.ctx import LeavePrbAction
 from gui.shared import EVENT_BUS_SCOPE, events, g_eventBus
 from helpers import dependency
 from skeletons.gui.lobby_context import ILobbyContext
+RestrictedEvent = namedtuple('RestrictedEvent', ('name', 'scope'))
 
 class CloseConfirmatorsHelper(object):
-    __slots__ = ('__closeConfirmator',)
+    __slots__ = ('__closeConfirmator', '_restrictedUrls', '_restrictedEvents', '_restrictedSfViews', '_restrictedGuiImplViews')
     _lobbyContext = dependency.descriptor(ILobbyContext)
+    ADD_HEADER_NAV_CONFIRMATOR = True
 
     def __init__(self):
         self.__closeConfirmator = None
+        self._restrictedUrls = set()
+        self._restrictedEvents = self.getRestrictedEvents()
+        self._restrictedSfViews = self.getRestrictedSfViews()
+        self._restrictedGuiImplViews = self.getRestrictedGuiImplViews()
+        self.__parseRestrictedUrls()
         return
 
     def getRestrictedEvents(self):
-        return [events.ViewEventType.LOAD_VIEW,
-         events.ViewEventType.LOAD_GUI_IMPL_VIEW,
-         events.ReferralProgramEvent.SHOW_REFERRAL_PROGRAM_WINDOW,
-         events.RallyWindowEvent.ON_CLOSE,
-         events.PrbInvitesEvent.ACCEPT,
-         events.PrbActionEvent.SELECT,
-         events.PrbActionEvent.LEAVE,
-         events.TrainingEvent.RETURN_TO_TRAINING_ROOM,
-         events.TrainingEvent.SHOW_TRAINING_LIST,
-         events.CustomizationEvent.SHOW]
+        return [RestrictedEvent(events.ViewEventType.LOAD_VIEW, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.ViewEventType.LOAD_GUI_IMPL_VIEW, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.ReferralProgramEvent.SHOW_REFERRAL_PROGRAM_WINDOW, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.RallyWindowEvent.ON_CLOSE, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.PrbInvitesEvent.ACCEPT, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.PrbActionEvent.SELECT, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.PrbActionEvent.LEAVE, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.TrainingEvent.RETURN_TO_TRAINING_ROOM, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.TrainingEvent.SHOW_TRAINING_LIST, EVENT_BUS_SCOPE.LOBBY),
+         RestrictedEvent(events.CustomizationEvent.SHOW, EVENT_BUS_SCOPE.LOBBY)]
 
     def getRestrictedSfViews(self):
         return [VIEW_ALIAS.VEHICLE_COMPARE,
@@ -43,17 +52,22 @@ class CloseConfirmatorsHelper(object):
     def getRestrictedGuiImplViews(self):
         return [R.views.lobby.dog_tags.DogTagsView()]
 
+    def getRestrictedUrls(self):
+        return []
+
     def start(self, closeConfirmator):
         self.__closeConfirmator = closeConfirmator
-        self._lobbyContext.addHeaderNavigationConfirmator(self.__confirmHeaderNavigation)
-        for event in self.getRestrictedEvents():
-            g_eventBus.addRestriction(event, self.__confirmEvent, scope=EVENT_BUS_SCOPE.LOBBY)
+        if self.ADD_HEADER_NAV_CONFIRMATOR:
+            self._lobbyContext.addHeaderNavigationConfirmator(self.__confirmHeaderNavigation)
+        for event, scope in self._restrictedEvents:
+            g_eventBus.addRestriction(event, self.__confirmEvent, scope=scope)
 
     def stop(self):
         self.__closeConfirmator = None
-        self._lobbyContext.deleteHeaderNavigationConfirmator(self.__confirmHeaderNavigation)
-        for event in self.getRestrictedEvents():
-            g_eventBus.removeRestriction(event, self.__confirmEvent, scope=EVENT_BUS_SCOPE.LOBBY)
+        if self.ADD_HEADER_NAV_CONFIRMATOR:
+            self._lobbyContext.deleteHeaderNavigationConfirmator(self.__confirmHeaderNavigation)
+        for event, scope in self.getRestrictedEvents():
+            g_eventBus.removeRestriction(event, self.__confirmEvent, scope=scope)
 
         return
 
@@ -67,11 +81,11 @@ class CloseConfirmatorsHelper(object):
     @wg_async
     def __confirmEvent(self, event, callback):
         if event.eventType == events.ViewEventType.LOAD_VIEW:
-            if event.alias not in self.getRestrictedSfViews():
+            if event.alias not in self._restrictedSfViews:
                 callback(True)
                 return
         if event.eventType == events.ViewEventType.LOAD_GUI_IMPL_VIEW:
-            if event.alias not in self.getRestrictedGuiImplViews():
+            if event.alias not in self._restrictedGuiImplViews:
                 callback(True)
                 return
         if event.eventType == events.PrbActionEvent.LEAVE:
@@ -79,8 +93,18 @@ class CloseConfirmatorsHelper(object):
                 if event.action.ignoreConfirmation:
                     callback(True)
                     return
+        if event.eventType == events.BrowserEvent.BROWSER_CREATED:
+            if event.ctx.get('url') not in self._restrictedUrls:
+                callback(True)
+                return
         result = yield wg_await(self.__closeConfirmator())
         callback(result)
+
+    @adisp.adisp_process
+    def __parseRestrictedUrls(self):
+        for url in self.getRestrictedUrls():
+            parsedUrl = yield URLMacros().parse(url)
+            self._restrictedUrls.add(parsedUrl)
 
     @adisp.adisp_async
     @wg_async
